instagram_clone/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0
django-cloudinary-storage==0.3.0
Django-Verify-Email==1.0.9
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.9.3
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
